implement basic registering of actions.
As of now, actions are only checked against a hard-coded structure as
verification. In the future, it would be nice to abstract this
verification, or at least put the verification of each action type
into a separate method of the Registry class.

The Registry does nothing further with the action definitions.