rework json interface
in the new config file format, actions are stored as key-value pairs in
an object, instead of objects in a list. Keys are names of actions. The value
depends on the action.

I am still not pleased with this format, it seems a lot more messy than
the previous one. This is still a WIP.