clean up config_file_format.md
the goal was to remove as many inline code elements as possible.
Names of action types and parameters are now written in bold italic,
similar to the style of python libraries documentation (e.g. matplotlib).
Expected type and optionality of action parameters written next to
the parameter names, also in bold italic.

Inline code fontstyle is now only used for technical things such as
names of gamerules or files.